syscalls: _getfsys takes a uint16_t not an int.
From: Sergio L. Pascual <https://github.com/slp/FUZIX>

+++ Library/include/syscalls.h
@@ -128,7 +128,7 @@ extern int rename(const char *path, const char *newpath);
 extern int _getdirent(int fd, void *buf, int len);
 extern int _stat(const char *path, struct _uzistat *s);
 extern int _fstat(int fd, struct _uzistat *s);
-extern int _getfsys(int dev, struct _uzifilesys *fs);
+extern int _getfsys(uint16_t dev, struct _uzifilesys *fs);
 extern int _time(time_t *t, uint16_t clock);
 extern int _stime(const time_t *t, uint16_t clock);
 extern int _times(struct tms *t);
